requirements2.txt
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
kiwisolver==1.4.8
matplotlib==3.10.5
narwhals==2.0.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pip==25.2
plotly==6.2.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6_sip==13.10.2
pyqt6-tools==6.4.2.3.3
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
setuptools==65.5.1
six==1.17.0
wheel==0.45.1
